Add constructor for scale
