+++ src/main/java/com/databricks/jdbc/core/DatabricksResultSet.java
@@ -41,7 +41,7 @@ public class DatabricksResultSet implements ResultSet, IDatabricksResultSet {
     this.statementId = statementId;
     this.executionResult =
         ExecutionResultFactory.getResultSet(resultData, resultManifest, statementId, session);
-    this.resultSetMetaData = new DatabricksResultSetMetaData(statementId, resultManifest);
+    this.resultSetMetaData = new DatabricksResultSetMetaData(statementId, resultManifest, session);
     this.statementType = statementType;
     this.updateCount = null;
     this.parentStatement = parentStatement;
@@ -93,6 +93,34 @@ public class DatabricksResultSet implements ResultSet, IDatabricksResultSet {
             columnTypeText,
             columnTypes,
             columnTypePrecisions,
+            rows.size());
+    this.statementType = statementType;
+    this.updateCount = null;
+    this.parentStatement = null;
+    this.isClosed = false;
+  }
+
+  public DatabricksResultSet(
+      StatementStatus statementStatus,
+      String statementId,
+      List<String> columnNames,
+      List<String> columnTypeText,
+      List<Integer> columnTypes,
+      List<Integer> columnTypePrecisions,
+      List<Integer> columnTypeScale,
+      List<List<Object>> rows,
+      StatementType statementType) {
+    this.statementStatus = statementStatus;
+    this.statementId = statementId;
+    this.executionResult = ExecutionResultFactory.getResultSet(rows);
+    this.resultSetMetaData =
+        new DatabricksResultSetMetaData(
+            statementId,
+            columnNames,
+            columnTypeText,
+            columnTypes,
+            columnTypePrecisions,
+            columnTypeScale,
             rows.size());
     this.statementType = statementType;
     this.updateCount = null;
